Do internal cleanup
